Back to float, cpu embedding

+++ GraphRec-WWW19/Attention.py
@@ -6,10 +6,10 @@ class Attention(nn.Module):
     def __init__(self, embedding_dims):
         super(Attention, self).__init__()
         self.embed_dim = embedding_dims
-        self.bilinear = nn.Bilinear(self.embed_dim, self.embed_dim, 1).half()
-        self.att1 = nn.Linear(self.embed_dim * 2, self.embed_dim).half()
-        self.att2 = nn.Linear(self.embed_dim, self.embed_dim).half()
-        self.att3 = nn.Linear(self.embed_dim, 1).half()
+        self.bilinear = nn.Bilinear(self.embed_dim, self.embed_dim, 1)
+        self.att1 = nn.Linear(self.embed_dim * 2, self.embed_dim)
+        self.att2 = nn.Linear(self.embed_dim, self.embed_dim)
+        self.att3 = nn.Linear(self.embed_dim, 1)
         self.softmax = nn.Softmax(0)
 
     def forward(self, node1, u_rep, num_neighs):
